Improved score_pdb.

+++ 1_dock/score_pdb.py
@@ -12,7 +12,7 @@ from containers import Protein
 from statistics import statistics
 from glob import glob
 
-alpha_factors = [0.5, 1.0, 1.5, 2.0]
+alpha_factors = [0.5, 1.0, 1.5, 2.0, 2.5]
 features = [['mcss', 'pipi', 'contact', 'hbond', 'sb'],
             ['mcss', 'contact', 'hbond', 'sb'],
             ['pipi', 'contact', 'hbond', 'sb'],
@@ -23,7 +23,6 @@ features = [['mcss', 'pipi', 'contact', 'hbond', 'sb'],
 
 cmd = '$SCHRODINGER/run {0:}/3_analyze/scores.py {1:} {1:} {1:} {1:}'.format(
                                                   shared_paths['code'], '{}')
-
 settings = {
     'num_poses' : 100,
     'chembl': False,
@@ -32,13 +31,22 @@ settings = {
 
 def score(stats_root, struct, protein, ligands, use_crystal_pose,
           alpha_factor, features, crystal = False):
-    settings['alpha'] = float(len(ligands)) * alpha_factor
+    subdir = '{}-{}'.format(alpha_factor, '_'.join(features))
+    if os.path.exists(subdir):
+        if not glob('{}/*.sc'.format(subdir)):
+            print('Missing output in {}/{}.'.format(os.getcwd(), subdir))
+        return
+    os.mkdir(subdir)
+    os.chdir(subdir)
+    
+    if crystal:
+        settings['alpha'] = alpha_factor
+    else:
+        settings['alpha'] = float(len(ligands)-1) * alpha_factor
     settings['use_crystal_pose'] = use_crystal_pose
     settings['k_list'] = features
 
-    subdir = '{}-{}'.format(alpha_factor, '_'.join(features))
-    os.mkdir(subdir)
-    os.chdir(subdir)
+
     
     with open('settings.py', 'w') as f:
         for varname, var in settings.items():
@@ -53,7 +61,7 @@ def score(stats_root, struct, protein, ligands, use_crystal_pose,
                 f.write(cmd.format(stats_root, struct, protein, ligand)+'\n')
         else:
             f.write(cmd.format(stats_root, struct, protein, ' '.join(ligands))+'\n')
-    os.system('sbatch -t 00:30:00 -p owners run.sh')
+    os.system('sbatch -t 01:00:00 -p owners run.sh')
     os.chdir('..')
 
 print('There are {} total jobs.'.format(  len(alpha_factors)
@@ -70,15 +78,13 @@ def run():
                                  shared_paths['stats']['version'])
         stats_root = '{}/{}'.format(root, 'stats')
         pdb_root = '{}/{}'.format(root, 'pdb')
-        if os.path.exists(pdb_root):
-            print('Already ran:', d)
-            continue
+
         os.system('mkdir -p {}'.format(root))
-        os.system('mkdir {}'.format(stats_root))
-        os.system('mkdir {}'.format(pdb_root))
-        os.system('mkdir {}/{}'.format(pdb_root, 'standard'))
-        os.system('mkdir {}/{}'.format(pdb_root, 'crystal'))
-        os.system('mkdir {}/{}'.format(pdb_root, 'only_crystal'))
+        os.system('mkdir -p {}'.format(stats_root))
+        os.system('mkdir -p {}'.format(pdb_root))
+        os.system('mkdir -p {}/{}'.format(pdb_root, 'standard'))
+        os.system('mkdir -p {}/{}'.format(pdb_root, 'crystal'))
+        os.system('mkdir -p {}/{}'.format(pdb_root, 'only_crystal'))
 
         protein = Protein(d)
         ligands = protein.get_xdocked_ligands(20)
@@ -117,38 +123,47 @@ def run():
         os.chdir('..')
 
 def check():
-    template = '{}/*/scores/{}/*/*/*/slurm*'.format(shared_paths['data'],
+
+    template = '{}/*/scores/{}/*/*/*/'.format(shared_paths['data'],
                                                   shared_paths['stats']['version'])
     print(template)
-    i, total = 0, 0
-    for fname in glob(template):
+    errors = []
+    total = 0
+    for fname in glob(template+'slurm*'):
         with open(fname) as fp:
             text = fp.read()
-            if 'Error' in text or 'Exception' in text:
-                print(fname)
-                print(text)
-                i += 1
+            if 'Error' in text or 'Exception' in text or 'error' in text:
+                if len(errors) < 10:
+                    # Don't write too many error messages
+                    print(fname)
+                    print(text)
+                errors += ['/'.join(fname.split('/')[:-1])]
         total += 1
-        if i > 10:
-            # Don't write too many error messages
-            exit()
-    print(total)
+    
+    print('{} slurm files'.format(total))
+    print('{} scoring scripts'.format(len(glob(template+'run.sh'))))
+    print(' '.join(errors))
 
 def merge():
     template = '{}/*/scores/{}/pdb/*/*/*.sc'.format(shared_paths['data'],
                                                     shared_paths['stats']['version'])
 
-    out_fname = '{}/../bpp_scores/{}.tsv'.format(shared_paths['data'],
-                                                 shared_paths['stats']['version'])
+    out_fname = '{}/../bpp_outputs/{}.tsv'.format(shared_paths['data'],
+                                                  shared_paths['stats']['version'])
     total = 0
     with open(out_fname, 'w') as out:
+        out.write('\t'.join(['mode', 'protein', 'ligand', 'alpha', 'features',
+                             'combind_rank', 'combind_rmsd',
+                             'glide_rank',   'glide_rmsd',
+                             'best_rank',    'best_rmsd']) + '\n')
         for fname in glob(template):
-            protein, _, version, pdb, mode, settings, name = template.split('/')[-6:]
-            settings = settings[:-3].split('-')
+            protein, _, version, pdb, mode, settings, name = fname.split('/')[-7:]
+            settings = settings.split('-')
             assert len(settings) == 2
             assert pdb == 'pdb'
             assert version == shared_paths['stats']['version']
             with open(fname) as fp:
+                fp.readline()
                 for line in fp:
                     tok = line.strip().split(',')
                     if len(tok) == 3:
@@ -163,6 +178,7 @@ def merge():
                         out.write('\t'.join([
                                             mode,
                                             protein,
+                                            lig,
                                             settings[0],
                                             settings[1],
                                             combind_rank, combind_rmsd,
@@ -170,7 +186,7 @@ def merge():
                                             best_rank, best_rmsd
                                             ]) + '\n')
                         total += 1
-    print('Processed a total of {} entries.'.format())
+    print('Processed a total of {} entries.'.format(total))
 
 def archive():
     template = '{}/*/scores/{}'.format(shared_paths['data'],
@@ -182,12 +198,31 @@ def archive():
         return
 
     for path in glob(template):
-        os.system('tar -zcvf {1:}.tar.gz {1:}'.format(path))
+        if os.path.exists(path + '.tar.gz'):
+            print(path)
+            continue
+        os.system('tar -zcvf {0:}.tar.gz {0:}'.format(path))
         os.system('rm -rd {}'.format(path))
+
+def inflate():
+    template = '{}/*/scores/{}.tar.gz'.format(shared_paths['data'],
+                                              shared_paths['stats']['version'])
+    paths = glob(template)
+    for path in paths:
+        print(path)
+    if input('Inflate all of the above tarballs? (yes)') != 'yes':
+        return
+
+    for path in glob(template):
+        if os.path.exists(path.replace('.tar.gz', '')):
+            print(path)
+            continue
+        os.system('tar -xvzf {}.tar.gz'.format(path))
+        os.system('rm {}'.format(path))
 
 if __name__ ==  '__main__':
     import sys
-    if sys.argv[1] == 'run':
+    if   sys.argv[1] == 'run':
         run()
     elif sys.argv[1] == 'check':
         check()
@@ -195,5 +230,7 @@ if __name__ ==  '__main__':
         merge()
     elif sys.argv[1] == 'archive':
         archive()
+    elif sys.argv[1] == 'inflate':
+        inflate()
     else:
         assert False
